Adding Pages
There are outlines of pages already in the index.html. I used these and created more pages. There are 4 given questions and I will have to come up with 2 new questions (one for allergy on page 5 and one for cold on page 6). I also added a results page (Page 7) at the end.