Added Yes/No Buttons for Question/page 1
The way to add buttons is given in the app assignment. When the user clicks either Yes or No to the button, the allergy count or cold count should change. I will test this after this commit and then continue by copying and pasting the button code to all pages.