Buttons added to 6 pages and Question 5/6 updated
The code from Page 1 was recycled to all questions/pages. Pages 5 and 6 have questions that were created based on information from the following website: https://www.bannerhealth.com/healthcareblog/better-me/how-to-tell-the-differences-between-allergies-and-colds. This website was also added to the top comment area (below "Known Issues"). Buttons for all pages will be tested after this commit.