Button Update and Page Reference Update
When running code on the Page link, I press the button and it sends me back to the "Get Started" page. This must mean my buttons need to be redone. Upon looking at my buttons, it looks like the reference page sends it back to two instead of gradually going up. All reference pages should send to the next page upon the press. This also means I needed to update the page IDs to correspond with the page number. These updates were performed on all pages and all buttons.