Create getResults when button is pressed on Results Page
Rather than coding the algorithm for the Results Page on the actual page, JavaScript and HTML code should be kept separate. Thus, a "getResult" method was implemented in JavaScript (between the <script> ... </script>) on line 60. I need to figure out the code within the method, so for now it will just say "This is where the results will be displayed."

On line 212 and 213, a button was created using advice from stackoverflow.