Revised getResult method & button on Results Page
Using the "App 3 Assignment -CorA 2023" document, I was able to add to the getResult method in JavaScript. After that, I created a button on the Results page (line 220). When the user clicks the button, the code should call the getResult method and will report the allergy percentage and the cold percentage.